+++ laravel/app/DocumentModels/Cpf.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Documents;
+namespace App\DocumentModels;
 
 use App\Concepts\MaskedDocument;
 
-class Cnpj extends MaskedDocument
+class Cpf extends MaskedDocument
 {
     /**
      * @return bool
@@ -12,14 +12,16 @@ class Cnpj extends MaskedDocument
     private function isFirstDigitValid(): bool
     {
         $sum = 0;
-        for ($index = 0; $index < 4; $index++) {
-            $sum += (5 - $index) * ($this->value[$index] + $this->value[8 + $index]) + ((9 - $index) * $this->value[4 + $index]);
+        $firstMultiplier = 10;
+        for ($index = 0; $index <= 8; $index++) {
+            $sum += $this->value[$index] * ($firstMultiplier - $index);
         }
-        $calculatedDigit = 11 - ($sum % 11);
-        if ($calculatedDigit >= 10) {
+        if (($sum % 11) < 2) {
             $calculatedDigit = 0;
+        } else {
+            $calculatedDigit = 11 - ($sum % 11);
         }
-        return $calculatedDigit == $this->value[12];
+        return $calculatedDigit == $this->value[9];
     }
 
     /**
@@ -28,18 +30,19 @@ class Cnpj extends MaskedDocument
     private function isSecondDigitValid(): bool
     {
         $sum = 0;
-        for ($index = 0; $index < 4; $index++) {
-            $sum += (6 - $index) * ($this->value[$index] + $this->value[8 + $index]);
+        $firstMultiplier = 11;
+        for ($index = 0; $index <= 9; $index++) {
+            if (str_repeat($index, 11) == $this->value) {
+                return false;
+            }
+            $sum += $this->value[$index] * ($firstMultiplier - $index);
         }
-        $sum += 2 * ($this->value[4] + $this->value[12]);
-        for ($index = 5; $index < 8; $index++) {
-            $sum += (14 - $index) * $this->value[$index];
-        }
-        $calculatedDigit = 11 - ($sum % 11);
-        if ($calculatedDigit >= 10) {
+        if (($sum % 11) < 2) {
             $calculatedDigit = 0;
+        } else {
+            $calculatedDigit = 11 - ($sum % 11);
         }
-        return $calculatedDigit == $this->value[13];
+        return $calculatedDigit == $this->value[10];
     }
 
     /**
@@ -47,7 +50,7 @@ class Cnpj extends MaskedDocument
      */
     public function isValid(): bool
     {
-        return strlen($this->value) == 14
+        return strlen($this->value) == 11
             && $this->isFirstDigitValid()
             && $this->isSecondDigitValid();
     }
@@ -59,8 +62,8 @@ class Cnpj extends MaskedDocument
     {
         $value = $this->getUnmaskedValue();
         if ($value) {
-            $pattern = "/^(\d{2})(\d{3})(\d{3})(\d{4})(\d{2})$/";
-            $replacement = "$1.$2/$3-$4";
+            $pattern = "/^(\d{3})(\d{3})(\d{3})(\d{2})$/";
+            $replacement = "$1.$2.$3-$4";
             return preg_replace($pattern, $replacement, $value);
         }
         return null;
